[1/2] base: volume key music control
Forward port from 4.4 times
Big thanks to CM for some important hints how to
implement this e.g.
https://github.com/CyanogenMod/android_frameworks_base/commit/e9a5363f51d22e776813ae18272734dfc992947f

Change-Id: Ib0839d4fb7f411d85023aa1de61716b7f78f2e97

ambient display: Fix volume key music control

Previously the dream service would only report whether or not it was dreaming.
Ambient display counts as dreaming.  When dreaming, the volume inputs are
passed back to the application which is not what we want for ambient
display.

Fix this by exposing isDozing() and allow volume keys to control music
when we are dreaming and have moved into doze mode (aka, sleeping).

Change-Id: I3b1e72b95ec8b1b63ef0219259dfee77a5d339d3
Signed-off-by: klozz <klozz707@gmail.com>