Fix clipping for edit screen of Quick Settings
The width of the edit screen is fixed when the display size is updated
in landscape mode. So when the device is rotated to portrait mode, the
width is not updated and the edit screen is truncated.
To fix this issue, "match_parent" should be used to match to the width
of Notification Panel always.

Bug: 34206618
Test: manual - go to Settings, rotate device to landscape mode, change
               display size (Default > Large > Default), rotate device
               to portrait mode and show QS edit screen

Change-Id: I35ac623812908a61cd34357e6bda567cab55ace1