base: Introduce Accidental Touch
Accidental Touch is a feature useful for many, especially the gamers out there.
It prevents any sort of accidental touch on hardware buttons while the touchscreen is
in use.
This means you won’t experience any unexpected behaviour while playing games,
web browsing or even when attempting to reach your finger out to the edge of
that humongous display.

Change-Id: Iee0ee1ad2913471447e20238cba3179084fa758a
Signed-off-by: klozz <shaclone707@gmail.com>