Disable haptic feedback on power menu items without long press actions
Power menu items without long press actions should not have haptic feedback.
It not only makes zero sense but gives the user the impression that something
is not working correctly. Haptic feedback is generally used to notify the user
that an action has been triggered. If there's no action to trigger than there
should not be haptic feedback.

- Reworked screenshot action to mirror the Reboot action in order to make this
play nice :-)

Change-Id: Ic935833c5990115378b2fa9d9a6f5a8aac81fc01
Signed-off-by: klozz <shaclone707@gmail.com>