Fix themes with the following squash of commits =============================================== Theme the resolver
Thanks to ezio84 to letting me know about the issue here. The issue being
that the theme resolver while not themed was good enough but if you used a
white accent you ended up with white text on a white background.

By themeing it, this issue is resolved :-)
==============================================
Theme immersive cling
=============================================
Also rename some themes

Change-Id: Id8beedec024e3af807be99254288cf4e0f08c5a9
Signed-off-by: klozz <klozz707@gmail.com>