Music ticker: once per track is enough
Remember the track that last was ticker-ed, and only ticker again
if actually something changed, instead of spamming with tickers
so that the status bar gets unusable.

--------

abc/du edit: keep previous track infos when pausing the music
so playing it again doesn't tick the same infos.
reset saved infos when dropping the media session (closing the player)
so a new player session can tick track infos even if it's the same
track

Change-Id: Ie91c4e75cffd558d12c83096f0f31bc40968875c
Signed-off-by: klozz <klozz707@gmail.com>