Add config to disallow the battery % option
This config will be respected and override the value found in Settings
if a user restores from a device which had SHOW_BATTERY_PERCENTAGE set
to true. It is also respceted on the keyguard where we would typically
show the battery percent if the device is charging.

Percentage is still shown above QS though.

Test: visual
Change-Id: I1f59b68097ffcbb7dda9d2ec99ea0ba9062c6dff
Bug: 80129194
Signed-off-by: klozz <klozz@vivaldi.net>