Add special captive portal servers for chinese
As we all know,google.com has been banned in China. So, we who are in China can't
connect to the CaptivePortal server. This causesa  problem where All networks to include
wifi and mobile data, are mistakenly detected as having no network connection.

This solution sets the appropriate server by detecting the system default language

*XPerience Edits*
- Use our Util class to avoid having to add in extra method

Change-Id: Ie5c9069050bbdc2aaf8587dbd61e7ae9aee04dc4
Signed-off-by: klozz <shaclone707@gmail.com>