Don't tell Assistant to hide on gesture completion or cancel
hide() is called by DefaultUiController itself when a gesture
fails to complete, or when the gesture animation finishes. (see
DefaultUiController.animateInvocationCompletion() and progress == 0
branch in DefaultUiController.onInvocationProgress())

hide() calls AssistManager.hideAssist(), which ultimately calls
VoiceInteractionSession#onCloseSystemDialogs(). We don't intend to
ask the assist app to close its UI when the gesture completes or when
invocation progress == 0.

Specifically, NgaUiController.hide /does not/ actually hide the
assistant (just our UI), and therefore
[DefaultUiController|GoogleDefaultUiController] shouldn't either. There
are times where NgaUiController does tell the assistant to close,
but they're specific to the NGA UI (e.g. there was a tap outside).

Bug: 136153254
Test: Verify WAI
Change-Id: I1cb9cb28a1009808e22571b2658f50f57b3e25b4